Move open_yaml and required_yaml_key to ssgcommon
